requirements.txt
pandas==2.1.0
numpy==1.24.3
geopy==2.4.0
requests==2.31.0
python-dotenv==1.0.0